pyproject.toml
[tool.poetry]
name = "brewblox-tilt"
version = "2.1.0"
description = "Tilt hydrometer service for Brewblox"
authors = ["BrewPi <development@brewpi.com>", "James Sandford <brewblox-tilt@j616s.co.uk>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
brewblox-service = "~1.0"
numpy = "1.21.5"
Pint = "^0.17"
PyBluez = "0.22"
beacontools = "^2.1.0"
"ruamel.yaml" = "^0.17.17"

[tool.poetry.dev-dependencies]
pytest-cov = "*"
pytest-mock = "*"
pytest-aiohttp = "*"
flake8 = "*"
autopep8 = "*"
flake8-quotes = "*"
click = "*"
PyYAML = "*"
pytest = "*"
aresponses = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
